refactor: moved code for getting filename into a separate block
